Fixed Docstrings
